Extend templating to allow custom initializer lists
Sometimes it is desirable to add members to the classes which
would require to extend the initializer lists.
At the moment, this cannot be done since no marked region
exists which will be migrated when updating the module.

So lets add such a region. Nitpick is, that I did not find
a nice way to handle the comma after the pre-generated list.
So this is left to the module author to take care - at least
we can give a hint in the description.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>