Create script to remove orphaned object / gcno files and all gcda files
Move coverage script to ev dev tools

Rename coverage tool to ev-coverage

Update ev coverage tool so it can be expanded easily. Make some ifs / for loops shorter. Add .idea to .gitignore.

Add version information.

Add "rm" as alias for remove_files

Prevent crash when source file cannot be found in current source tree

This seems to be the case for object files in _deps subdirs where the corresponding source code location isn't immediately obvious to this tool
At the moment these .o file are removed as well but this may not be desired behavior

Added --dry-run
Added --silent to suppress all output
Added --summary to print a summary of the number of removed/to-be-removed files

Use dwarfdump or gdb to find the correct cpp file that belongs to the object file.

Fix dwarfdump result where a list of multiple files is returned instead of one file.

Signed-off-by: Maaike Zijderveld, iolar <git.mail@iolar.nl>
Signed-off-by: Kai-Uwe Hermann <kai-uwe.hermann@pionix.de>